refactor: split oversized router.rs and engine.rs into modules
- harness-server: router.rs (1852L) → router/mod.rs (201L) + router/tests.rs (1650L)
- harness-rules: engine.rs (1388L) → engine/mod.rs (777L) + engine/tests.rs (610L)
- fix pre-existing TaskState initializers in services/task.rs missing fields

Closes #227
